Marks y nueva build

+++ DataBase/tfg/src/main/java/com/tfg_data_base/tfg/Marks/Mark.java
@@ -10,7 +10,7 @@ public class Mark {
     @Id
     private String id;
     private String name;
-    private Long x;
-    private Long y;
+    private Double x;
+    private Double y;
 
 }
